Create a shared base class for AddView
