+++ wagtail/documents/views/multiple.py
@@ -5,11 +5,9 @@ from django.http import HttpResponseBadRequest, JsonResponse
 from django.shortcuts import get_object_or_404
 from django.template.loader import render_to_string
 from django.urls import reverse
-from django.utils.decorators import method_decorator
-from django.views.decorators.vary import vary_on_headers
-from django.views.generic import TemplateView, View
+from django.views.generic import View
 
-from wagtail.admin.views.generic import PermissionCheckedMixin
+from wagtail.admin.views.generic.multiple_upload import AddView as BaseAddView
 from wagtail.search.backends import get_search_backends
 
 from .. import get_document_model
@@ -18,9 +16,8 @@ from ..models import UploadedDocument
 from ..permissions import permission_policy
 
 
-class AddView(PermissionCheckedMixin, TemplateView):
+class AddView(BaseAddView):
     permission_policy = permission_policy
-    permission_required = 'add'
     template_name = 'wagtaildocs/multiple/add.html'
     edit_form_template_name = 'wagtaildocs/multiple/edit_form.html'
     upload_model = UploadedDocument
@@ -33,12 +30,6 @@ class AddView(PermissionCheckedMixin, TemplateView):
 
     def get_edit_form_class(self):
         return get_document_multi_form(self.model)
-
-    @method_decorator(vary_on_headers('X-Requested-With'))
-    def dispatch(self, request):
-        self.model = get_document_model()
-
-        return super().dispatch(request)
 
     def save_object(self, form):
         doc = form.save(commit=False)
@@ -83,15 +74,6 @@ class AddView(PermissionCheckedMixin, TemplateView):
             ),
         }
 
-    def get_invalid_response_data(self, form):
-        """
-        Return the JSON response data for an invalid form submission
-        """
-        {
-            'success': False,
-            'error_message': '\n'.join(form.errors['file']),
-        }
-
     def get_edit_upload_form_context_data(self):
         """
         Return the context data necessary for rendering the HTML form for supplying the
@@ -121,67 +103,6 @@ class AddView(PermissionCheckedMixin, TemplateView):
                 request=self.request
             ),
         }
-
-    def post(self, request):
-        if not request.is_ajax():
-            return HttpResponseBadRequest("Cannot POST to this view without AJAX")
-
-        if not request.FILES:
-            return HttpResponseBadRequest("Must upload a file")
-
-        # Build a form for validation
-        self.upload_form_class = self.get_upload_form_class()
-        form = self.upload_form_class({
-            'title': request.FILES['files[]'].name,
-            'collection': request.POST.get('collection'),
-        }, {
-            'file': request.FILES['files[]']
-        }, user=request.user)
-
-        if form.is_valid():
-            # Save it
-            self.object = self.save_object(form)
-
-            # Success! Send back an edit form for this document to the user
-            return JsonResponse(self.get_edit_object_response_data())
-        elif 'file' in form.errors:
-            # The uploaded file is invalid; reject it now
-            return JsonResponse(self.get_invalid_response_data())
-        else:
-            # Some other field of the document form has failed validation, e.g. a required metadata
-            # field on a custom document model. Store the document as an UploadedDocument instead
-            # and present the edit form so that it will become a proper Document when successfully
-            # filled in
-            self.upload_object = self.upload_model.objects.create(
-                file=self.request.FILES['files[]'], uploaded_by_user=self.request.user
-            )
-            self.object = self.model(
-                title=self.request.FILES['files[]'].name,
-                collection_id=self.request.POST.get('collection')
-            )
-
-            return JsonResponse(self.get_edit_upload_response_data())
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # Instantiate a dummy copy of the form that we can retrieve validation messages and media from;
-        # actual rendering of forms will happen on AJAX POST rather than here
-        self.upload_form_class = self.get_upload_form_class()
-        form = self.upload_form_class(user=self.request.user)
-
-        collections = self.permission_policy.collections_user_has_permission_for(self.request.user, 'add')
-        if len(collections) < 2:
-            # no need to show a collections chooser
-            collections = None
-
-        context.update({
-            'help_text': form.fields['file'].help_text,
-            'collections': collections,
-            'form_media': form.media,
-        })
-
-        return context
 
 
 class EditView(View):
